🚚 move mock data, cookie.jsx and rename util folder
